requirements.txt
anytree==2.4.3
appdirs==1.4.3
arabic-reshaper==2.0.14
argh==0.26.2
asn1crypto==0.24.0
astroid==2.1.0
attrs==18.2.0
black==18.9b0
cachetools==3.0.0
cachier==1.2.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.7.1
cryptography==2.3.1
cycler==0.10.0
dash==0.35.1
dash-bootstrap-components==0.3.1
dash-core-components==0.42.1
dash-html-components==0.13.4
dash-renderer==0.17.0
dash-table==3.1.11
decorator==4.3.2
Flask==1.0.2
Flask-Compress==1.4.0
future==0.17.1
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.0a3
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mkl-fft==1.0.10
mkl-random==1.0.2
nbformat==4.4.0
networkx==2.2
numpy==1.15.4
oauthlib==3.0.1
pandas==0.24.0
pathtools==0.1.2
plotly==3.5.0
portalocker==1.3.0
pyarrow==0.12.0a0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycosat==0.6.3
pycparser==2.19
pygsheets==2.0.0
pylint==2.2.2
pyOpenSSL==18.0.0
pyparsing==2.3.1
pyrsistent==0.14.9
PySocks==1.6.8
python-bidi==0.4.0
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.15.46
six==1.12.0
toml==0.10.0
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
watchdog==0.9.0
Werkzeug==0.14.1
wrapt==1.11.0
